--- models/Admin.js
@@ -1,43 +1,0 @@
-const { Model, DataTypes } = require('sequelize');
-const bcrypt = require('bcrypt');
-const sequelize = require('../config/connection.js');
-
-class Admin extends Model {
-      // set up method to run on instance data (per user) to check password
-      checkPassword(loginPw) {
-        return bcrypt.compareSync(loginPw, this.password);
-      }
-}
-
-Admin.init(
-  {
-    // define columns
-      username: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    password: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: {
-        len: [8]
-      }
-    },
-  },
-  {
-    hooks: {
-      // set up beforeCreate lifecycle "hook" functionality
-      beforeCreate: async (newUserData) => {
-        newUserData.password = await bcrypt.hash(newUserData.password, 10);
-        return newUserData;
-      },
-      beforeUpdate: async (updatedUserData) => {
-        updatedUserData.password = await bcrypt.hash(updatedUserData.password, 10);
-        return updatedUserData;
-      }
-    },
-    sequelize,
-  }
-);
-
-module.exports = Admin;
