+++ server.js
@@ -4,7 +4,7 @@ const sequelize = require("./config/connection");
 const session = require("express-session");
 const SequelizeStore = require("connect-session-sequelize")(session.Store);
 const exphbs = require("express-handlebars");
-const nodemailer = require("nodemailer");
+
 
 // Sets up the Express App
 // =============================================================
@@ -37,42 +37,6 @@ app.use(
 
 app.use("/", allRoutes);
 
-// app.get("/images", (req, res) => {
-//   res.render("images");
-// });
-
-// nodemailer==========================================================
-// let transporter = nodemailer.createTransport(transport[ defaults])
-// transporter.sendMail(data[ callback])
-app.post("/", async (req, res) => {
-  const { email } = req.body;
-  let transporter = nodemailer.createTransport({
-    host: "smtp.ethereal.email",
-    port: 587,
-    secure: false, // true for 465, false for other ports
-    auth: {
-      user: "gaston.shanahan66@ethereal.email", //testAccount.user, // generated ethereal user
-      pass: "YtnW7qgxNvjsPsXTpj", //testAccount.pass, // generated ethereal password
-    },
-  });
-
-  // send mail with defined transport object
-  const msg = {
-    from: "gaston.shanahan66@ethereal.email", //'need a tester email', // sender address
-    to: `gaston.shanahan66@ethereal.email`, // list of receivers
-    subject: "A friendly reminder", // Subject line
-    text: "The specials are", // plain text body
-    html: "", // html body
-  };
-  const info = await transporter.sendMail(msg);
-
-  console.log("Message sent: %s", info.messageId);
-  // Message sent: <b658f8ca-6296-ccf4-8306-87d57a0b4321@example.com>
-
-  // Preview only available when sending through an Ethereal account
-  res.send("Email Sent!");
-});
-// ========================================================================
 
 sequelize.sync({ force: false }).then(function () {
   app.listen(PORT, function () {
